requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
garminconnect==0.2.7
garth==0.4.47
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-telegram-bot==22.5
readchar==4.2.1
requests==2.32.5
requests-oauthlib==1.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
